refactor(boundary): share TypeScript source collection utility
