+++ src/typescript-file-collector.ts
@@ -1,0 +1,15 @@
+import { readdirSync, statSync } from "node:fs";
+import { resolve } from "node:path";
+
+export function collectTypeScriptFiles(dir: string): string[] {
+  return readdirSync(dir).flatMap((entry) => {
+    const fullPath = resolve(dir, entry);
+    const stat = statSync(fullPath);
+
+    if (stat.isDirectory()) {
+      return collectTypeScriptFiles(fullPath);
+    }
+
+    return fullPath.endsWith(".ts") ? [fullPath] : [];
+  });
+}
